move some config to dotfile

+++ setup.sh
@@ -1,22 +1,4 @@
 #!/bin/bash
-sudo apt-get update && sudo apt-get upgrade -y
-sudo apt-get install -y build-essential \
-    libncursesw5-dev \
-    libssl-dev \
-    libpcre3-dev \
-    libdb-dev \
-    zlib1g-dev \
-    libgdbm-dev \
-    liblzma-dev \
-    uuid-dev \
-    libicu-dev \
-    libxml2-dev \
-    libxslt1-dev \
-    libreadline-dev \
-    libevent-dev \
-    autoconf \
-    yodl \
-    gettext
 
 echo "======================="
 echo "Starting configuration setup..."
@@ -28,29 +10,17 @@ ZSHENV_FILE="$HOME/.zshenv"
 TMUX_CONF="$HOME/.tmux.conf"
 
 # Remove existing configuration files
-# echo "Removing old configurations..."
-# for file in "$CONFIG_DIR" "$ZSHENV_FILE" "$TMUX_CONF"; do
-#     if [ -e "$file" ]; then
-#         rm -rf "$file"
-#     fi
-# done
+echo "Removing old configurations..."
+for file in "$CONFIG_DIR" "$ZSHENV_FILE" "$TMUX_CONF"; do
+    if [ -e "$file" ]; then
+        rm -rf "$file"
+    fi
+done
 
 echo "Linking new configuration files..."
 ln -sf "$DOTFILES_DIR/config" "$CONFIG_DIR"
 ln -sf "$DOTFILES_DIR/zsh_config/zshenv" "$ZSHENV_FILE"
 ln -sf "$DOTFILES_DIR/tmux_config/tmux.conf" "$TMUX_CONF"
-
-# Install Anaconda
-echo "======================="
-echo "Installing Anaconda..."
-mkdir -p ~/.anaconda
-wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/.anaconda/miniconda.sh
-bash ~/.anaconda/miniconda.sh -b -u -p ~/.anaconda
-rm ~/.anaconda/miniconda.sh
-
-source $HOME/.anaconda/bin/activate && \
-conda activate && \
-pip install --upgrade pip
 
 echo "======================="
 echo "Setting zsh"
@@ -65,10 +35,14 @@ sudo mkdir -p /run/user/$USER_ID/fnm_multishells
 sudo chmod a+w /run/user/$USER_ID/fnm_multishells
 
 # thefuck
-# pip install https://github.com/nvbn/thefuck/archive/master.zip
+pip install https://github.com/nvbn/thefuck/archive/master.zip
 
 #starship
-conda install -c conda-forge starship --yes
+if [ -e "$HOME/.conda" ]; then
+    conda install -c conda-forge starship --yes
+else
+    curl -sS https://starship.rs/install.sh | sh -s -- -y
+fi
 
 wget https://www.zsh.org/pub/zsh-5.9.tar.xz -O zsh.tar.xz
 tar xf zsh.tar.xz
@@ -107,8 +81,5 @@ rm -rf "$GITSTATUSD_DIR/gitstatusd-linux-x86_64.tar.gz"
 typeset -g POWERLEVEL9K_GITSTATUS_DIR="$GITSTATUSD_DIR"
 
 zsh "$HOME/.dotfile/config/zsh/.zshrc"
-
-sudo apt-get -y install -y locales
-sudo locale-gen en_US.UTF-8
 
 echo "Base configuration setup completed successfully!"
